Avoid drilling into models module
