+++ enterprise/frontend/src/metabase-enterprise/content_verification/utils.ts
@@ -4,7 +4,7 @@ import { useEffect, useMemo, useState } from "react";
 import type {
   ActualModelFilters,
   AvailableModelFilters,
-} from "metabase/browse/models/utils";
+} from "metabase/browse/models";
 import { useUserSetting } from "metabase/common/hooks";
 
 export const availableModelFilters: AvailableModelFilters = {
