+++ frontend/src/metabase/browse/metrics/index.tsx
@@ -1,2 +1,6 @@
 export { BrowseMetrics } from "./BrowseMetrics";
-export type { MetricFilterControlsProps, MetricFilterSettings } from "./types";
+export type {
+  MetricFilterControlsProps,
+  MetricFilterSettings,
+  RecentMetric,
+} from "./types";
