+++ frontend/src/metabase/browse/models/index.tsx
@@ -1,1 +1,8 @@
 export { BrowseModels } from "./BrowseModels";
+export type {
+  AvailableModelFilters,
+  ActualModelFilters,
+  ModelFilterControlsProps,
+  ModelResult,
+  RecentModel,
+} from "./types";
