+++ frontend/src/metabase/browse/models/types.tsx
@@ -1,3 +1,5 @@
+import type { Dispatch, SetStateAction } from "react";
+
 import type { RecentCollectionItem, SearchResult } from "metabase-types/api";
 
 /** Model retrieved through the search endpoint */
@@ -12,3 +14,20 @@ export interface RecentModel extends RecentCollectionItem {
  * This type is needed so that our filtering functions can
  * filter arrays of models retrieved from either endpoint. */
 export type FilterableModel = ModelResult | RecentModel;
+
+export type AvailableModelFilters = Record<
+  string,
+  {
+    predicate: (value: FilterableModel) => boolean;
+    activeByDefault: boolean;
+  }
+>;
+
+export type ModelFilterControlsProps = {
+  actualModelFilters: ActualModelFilters;
+  setActualModelFilters: Dispatch<SetStateAction<ActualModelFilters>>;
+};
+
+/** Mapping of filter names to true if the filter is active
+ * or false if it is inactive */
+export type ActualModelFilters = Record<string, boolean>;
