+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbar.unit.spec.tsx
@@ -15,8 +15,8 @@ import {
   waitForLoaderToBeRemoved,
   within,
 } from "__support__/ui";
+import type { ModelResult } from "metabase/browse/models";
 import { createMockModelResult } from "metabase/browse/models/test-utils";
-import type { ModelResult } from "metabase/browse/models/types";
 import { ROOT_COLLECTION } from "metabase/entities/collections";
 import * as Urls from "metabase/lib/urls";
 import type { Card, Dashboard, DashboardId, User } from "metabase-types/api";
